powerpc/pseries: Call RTAS directly
The pseries guests call RTAS via a RTAS entry point which is a firmware
image under powernv and simple HCALL wrapper under QEMU. For the latter,
we can skip the binary image and do HCALL directly, eliminating the need
in the RTAS blob entirely.

This checks the DT whether the new method is supported and use it if it is.
This removes few checks as QEMU might decide not to keen RTAS around at
all (might be the case for secure VMs).

Note that kexec still check for the linux,rtas-xxx properties.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>