powernv/phb4: Do TCE invalidations directly via PHB MMIO
At the moment TCE cache is invalidated via OPAL which is costly as
it requires disabling MMU.

This implements direct access to the TCE Kill facility for real and
virtual modes (although real mode is unlikely to be called on POWER9).
If more than 1mil entries are to invalidate, this invalidates a PE cache.

This uses spinlock which has been added in 2011 but never been used
since then (this is why spin_lock_init() is missing here).

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>