KVM: PPC: Book3S HV: Stop returning internal values to userspace
Our kvmppc_vcpu_run_hv currently returns the RESUME_GUEST|HOST values
to userspace, against the API of the KVM_RUN ioctl which returns 0 on
success and -1 on error.

This was noticed because a subsequent patch that enables the kvm
selftests for powerpc hit an assert at the _vcpu_run function when we
return a value different from the expected.

Signed-off-by: Fabiano Rosas <farosas@linux.ibm.com>