coco/tsm: Add tsm and tsm-host modules
The modules responsibilities are:
1. bus agnostic interface for reading TDISP-capable devices
certificates/measurements/reports;
2. control for triggering device setup;
3. CoCo VM required ABI.

This relies on the platform to register a set of callbacks for the TSM,
for both host and guest.

A bus driver is expected to scan bus(es) and call this module to set up
sysfs nodes.

Add tdi_enabled in the device struct to allow direct DMA as otherwise
SWIOTLB or memory decryption will be enforced (used later in the series).

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>