iommufd: amd-iommu: Add vdevice support
The new AMD SEV-TIO feature allows DMA to/from encrypted memory and
encrypted MMIO for use in CoCo VMs (SEV-SNP). The secure part of IOMMU
is called sDTE ("Secure DTE") which resides in private memory and
controlled by the PSP. sDTEs of passed through devices are in the TCB
of CoCo VMs and inaccessible by the host OS.

Implement vdevice in the AMD IOMMU host OS to represent the host instance
of a secure IOMMU which is visible to the guest. Later on this is going to
be used for guest-to-PSP secure communication in order to configure sDTE
and encrypted MMIO.

Most parts of insecure DTE move to sDTE so DTEs need to be adjusted.
At the moment this includes "domain_id" (moves to sDTE) and
"IOTLB enable" (should stay in sync with sDTE).

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>
---

Skips =1 in iommu_viommu_type because there are ARM patches
floating around which define it and will get upstream first.

This "domain_id" is an AMD thing which needs to be unique.
The PSP is manages this for now.